+++ covasim/sim.py
@@ -208,7 +208,7 @@ class Sim(cvb.BaseSim):
         for key,label in cvd.result_stocks.items():
             self.results[f'n_{key}'] = init_res(label, color=dcols[key])
         self.results['n_susceptible'].scale = 'static'
-        self.results['bed_capacity']  = init_res('Percentage bed capacity', scale=False)
+        self.results['bed_capacity']  = init_res('Bed demand relative to capacity', scale=False)
 
         # Flows and cumulative flows
         for key,label in cvd.result_flows.items():
@@ -282,7 +282,7 @@ class Sim(cvb.BaseSim):
         inds = np.arange(int(self['pop_infected']))
         self.people.infect(inds=inds)
         for ind in inds:
-            self.people.transtree.seeds.append({'person':ind, 'date':self.t, 'layer':None})
+            self.people.transtree.linelist[ind] = dict(source=None, target=ind, date=self.t, layer='seed_infection')
 
         return
 
@@ -300,7 +300,7 @@ class Sim(cvb.BaseSim):
         # Perform initial operations
         self.rescale() # Check if we need to rescale
         people   = self.people # Shorten this for later use
-        flows    = people.update_states(t=t) # Update the state of everyone and count the flows
+        flows    = people.update_states_pre(t=t) # Update the state of everyone and count the flows
         contacts = people.update_contacts() # Compute new contacts
         bed_max  = people.count('severe') > self['n_beds'] # Check for a bed constraint
 
@@ -310,12 +310,26 @@ class Sim(cvb.BaseSim):
             imporation_inds = cvu.choose(max_n=len(people), n=n_imports)
             flows['new_infections'] += people.infect(inds=imporation_inds, bed_max=bed_max)
             for ind in imporation_inds:
-                self.people.transtree.seeds.append({'person':ind, 'date':self.t, 'layer':None})
+                self.people.transtree.linelist[ind] = dict(source=None, target=ind, date=self.t, layer='importation')
+
+        # Apply interventions
+        for intervention in self['interventions']:
+            intervention.apply(self)
+        if self['interv_func'] is not None: # Apply custom intervention function
+            self =self['interv_func'](self)
+
+        flows = people.update_states_post(flows) # Check for state changes after interventions
 
         # Compute the probability of transmission
         beta         = cvd.default_float(self['beta'])
         asymp_factor = cvd.default_float(self['asymp_factor'])
         diag_factor  = cvd.default_float(self['diag_factor'])
+        frac_time    = cvd.default_float(self['viral_dist']['frac_time'])
+        load_ratio   = cvd.default_float(self['viral_dist']['load_ratio'])
+        date_inf     = people.date_infectious
+        date_rec     = people.date_recovered
+        date_dead    = people.date_dead
+        viral_load = cvu.compute_viral_load(t, date_inf, date_rec, date_dead, frac_time, load_ratio)
 
         for lkey,layer in contacts.items():
             sources = layer['p1']
@@ -330,7 +344,7 @@ class Sim(cvb.BaseSim):
             quar       = people.quarantined
             quar_eff   = cvd.default_float(self['quar_eff'][lkey])
             beta_layer = cvd.default_float(self['beta_layer'][lkey])
-            rel_trans, rel_sus = cvu.compute_trans_sus(rel_trans, rel_sus, beta_layer, symp, diag, quar, asymp_factor, diag_factor, quar_eff)
+            rel_trans, rel_sus = cvu.compute_trans_sus(rel_trans, rel_sus, beta_layer, viral_load, symp, diag, quar, asymp_factor, diag_factor, quar_eff)
 
             # Calculate actual transmission
             target_inds, edge_inds = cvu.compute_infections(beta, sources, targets, betas, rel_trans, rel_sus) # Calculate transmission!
@@ -340,14 +354,9 @@ class Sim(cvb.BaseSim):
             for ind in edge_inds:
                 source = sources[ind]
                 target = targets[ind]
-                self.people.transtree.sources[target] = {'source':source, 'target':target, 'date':self.t, 'layer':lkey}
-                self.people.transtree.targets[source].append({'source':source, 'target':target, 'date':self.t, 'layer':lkey})
-
-        # Apply interventions
-        for intervention in self['interventions']:
-            intervention.apply(self)
-        if self['interv_func'] is not None: # Apply custom intervention function
-            self =self['interv_func'](self)
+                transdict = dict(source=source, target=target, date=self.t, layer=lkey)
+                self.people.transtree.linelist[target] = transdict
+                self.people.transtree.targets[source].append(transdict)
 
         # Update counts for this time step: stocks
         for key in cvd.result_stocks.keys():
@@ -356,8 +365,7 @@ class Sim(cvb.BaseSim):
 
         # Update counts for this time step: flows
         for key,count in flows.items():
-            if key != 'new_tests': # tests are updated separately, as part of interventions
-                self.results[key][t] = count
+            self.results[key][t] += count
 
         # Tidy up
         self.t += 1
@@ -588,7 +596,6 @@ class Sim(cvb.BaseSim):
         if self.data is None:
             return np.nan
 
-        pLowest = 1e-20
         loglike = 0
 
         model_dates = self.datevec.tolist()
@@ -635,7 +642,7 @@ class Sim(cvb.BaseSim):
     def plot(self, to_plot=None, do_save=None, fig_path=None, fig_args=None, plot_args=None,
              scatter_args=None, axis_args=None, legend_args=None, as_dates=True, dateformat=None,
              interval=None, n_cols=1, font_size=18, font_family=None, use_grid=True, use_commaticks=True,
-             do_show=True, verbose=None):
+             log_scale=False, do_show=True, verbose=None):
         '''
         Plot the results -- can supply arguments for both the figure and the plots.
 
@@ -656,6 +663,7 @@ class Sim(cvb.BaseSim):
             font_family (str): Font face
             use_grid (bool): Whether or not to plot gridlines
             use_commaticks (bool): Plot y-axis with commas rather than scientific notation
+            log_scale (bool or list): Whether or not to plot the y-axis with a log scale; if a list, panels to show as log
             do_show (bool): Whether or not to show the figure
             verbose (bool): Display a bit of extra information
 
@@ -690,6 +698,12 @@ class Sim(cvb.BaseSim):
         n_rows = np.ceil(len(to_plot)/n_cols) # Number of subplot rows to have
         for p,title,keylabels in to_plot.enumitems():
             ax = pl.subplot(n_rows, n_cols, p+1)
+            if log_scale:
+                if isinstance(log_scale, list):
+                    if title in log_scale:
+                        ax.set_yscale('log')
+                else:
+                    ax.set_yscale('log')
             for key in keylabels:
                 label = res[key].name
                 this_color = res[key].color