+++ tests/test_interventions_other.py
@@ -36,25 +36,20 @@ def test_beds(do_plot=False, do_show=True, do_save=False, fig_path=None):
           }
         },
         'bedconstraint': {
+            'name': 'Only 50 beds available',
+            'pars': {
+                'pop_infected': 100,
+                'n_beds': 50,
+            }
+        },
+        'bedconstraint2': {
             'name': 'Only 10 beds available',
             'pars': {
                 'pop_infected': 100,
                 'n_beds': 10,
             }
         },
-        'bedconstraint2': {
-            'name': 'Only 1 bed available',
-            'pars': {
-                'pop_infected': 100,
-                'n_beds': 1,
-                # 'OR_no_treat': 10., # nb. scenarios cannot currently overwrite nested parameters
-                # This prevents overwriting OR_no_treat due to recent refactoring but more generally
-                # there are other nested parameters eg. all of those under pars['dur']
-            }
-        },
     }
-
-
 
     scens = cv.Scenarios(sim=sim, basepars=basepars, metapars=metapars, scenarios=scenarios)
     scens.run(verbose=verbose, debug=debug)